rename blindutils to solverutils
