+++ include/astrometry/solverutils.h
