+++ solver/engine.c
@@ -28,7 +28,7 @@
 #include "solver.h"
 #include "math.h"
 #include "fitsioutils.h"
-#include "blindutils.h"
+#include "solverutils.h"
 #include "os-features.h"
 #include "onefield.h"
 #include "log.h"
