--- util/makefile.blind
@@ -1,5 +1,0 @@
-# This file is part of the Astrometry.net suite.
-# Licensed under a 3-clause BSD style license - see LICENSE
-
-SOLVER_DIR := $(BASEDIR)/solver
-SOLVER_INC := -I$(SOLVER_DIR)
